Product Form for specific user(s)

+++ shopping-app/routes/shopItemRoutes.js
@@ -4,12 +4,8 @@ models = sequelize.models;
 
 //Creating shop item
 routes.post("/create/shopItem", async function (req, res) {
-  if (req.body.id) {
-    res
-      .status(400)
-      .send(
-        `Bad request: ID should not be provided, since it is determined automatically by the database.`
-      );
+  if (req.body.id || req.body.user != "Scott50000@gmail.com") {
+    res.status(400).send(`Bad request. Are you not logged in as an admin?`);
   } else {
     await models.shopItem.create(req.body);
     res.status(201).end();
@@ -31,7 +27,13 @@ routes.get("/read/shopItem/:id", async function (req, res) {
 
 //Updating shop item
 routes.post("/update/shopItems/:id", async function (req, res) {
+  if (req.body.user == "Scott50000@gmail.com") {
+    res.status(400).send(`Not an admin.`);
+    return;
+  }
+
   const id = req.params.id;
+
   const shopItem = await models.shopItem.findByPk(id);
   if (req.body.id === id) {
     await models.shopItem.update(req.body, {
@@ -51,6 +53,11 @@ routes.post("/update/shopItems/:id", async function (req, res) {
 
 //Delete shop item
 routes.post("/delete/shopItem/:id", async function (req, res) {
+  if (req.body.user == "Scott50000@gmail.com") {
+    res.status(400).send(`Not an admin.`);
+    return;
+  }
+
   const id = req.params.id;
   await models.shopItem.destroy({
     where: {
